+++ Models/Customer2_rec.cs
@@ -7,23 +7,23 @@ namespace BuildsByBrickwellNew.Models;
 public class Customer2_rec
 {
     [Key]
-    public int? CustomerId { get; set; }
+    public byte? CustomerId { get; set; }
 
-    public int? Rec1 { get; set; }
+    public byte? Rec1 { get; set; }
 
-    public int? Rec2 { get; set; }
+    public byte? Rec2 { get; set; }
 
-    public int? Rec3 { get; set; }
+    public byte? Rec3 { get; set; }
 
-    public int? Rec4 { get; set; }
+    public byte? Rec4 { get; set; }
 
-    public int? Rec5 { get; set; }
+    public byte? Rec5 { get; set; }
 
-    public int? Rec6 { get; set; }
+    public byte? Rec6 { get; set; }
 
-    public int? Rec7 { get; set; }
+    public byte? Rec7 { get; set; }
 
-    public int? Rec8 { get; set; }
+    public byte? Rec8 { get; set; }
 
-    public int? Rec9 { get; set; }
+    public byte? Rec9 { get; set; }
 }
